requirements.txt
PyYAML~=6.0.1
setuptools==68.1.2
click~=8.1.7